+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+v1.1.0 (May 26 2021)
+===================
+  ### Added
+  * Support for separate command/payload input properties.
+  * Support for topic passthrough.
+  * Supports error catching with catch node.
+
+  ### Changed
+  * Removed msg.status property in lue of better error handling / catch node.
+  * Improved README.md
+  
 v1.0.2 / May 25 2021
 ===================